object state management in store for report info
